Made changes
Used webpack for more optimal and faster way for the app to load because it takes all the files we use for importing and creates only one final that needs to be run at the start. Also removed the use of the autobind decorator and i just use the build in .bind() method where needed. In addition removed the enum and its place added a projectStatus type that can have a value of the possible project statuses . Some other minor changes are the change of the name of the Listeners array to stateChangeListeners for easier understanding and made some changes to the comments in the project