Added a bin icon which on dropping an object removes it
Added a new Template in the index.html file for the bin icon also some css in the app.css file for better visualisation. In the components folder created a new file (project-bin.ts) which contains a new class ProjectBin that extends component and implements DragTarget. In the project state class i added a new method deleteProject which takes as a parameter the project id.We call the delete method in the dropHandler of the ProjectBin class after a confirmation from the user