Added ID And Parent ID fields and also when hovering over a projects parent Id some info about it appears
In the index.html file added 2 more input field to the form(Project Id and Parent Id) and in the project-item template added one more h3 which stores the parent Id and a div which when we hover over the parent Id appears and shows the title and the number of people about the project with that Id. In the project-input file added the appropriate validations for the Ids.In the project-item file made the appropriate changed and added mouseover and mouseout event listeners to the header which stores the parentId. In the project class added the extra variable parentId which is optional. In the Project State class added 2 new methods one to check if an Id already is taken and another one that return a project based on the given Id